feat(ui-scripts,ui-test-utils): add a --mocha option to ui-test
- remove the --node option
- add partial jsdom support to ui-test-utils

TEST PLAN:
- there should be no changes to the docs
- yarn start:examples should bring up storybook
- yarn test should pass

Reviewed-on: https://gerrit.instructure.com/181372
Tested-by: Jenkins
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>

+++ packages/ui-test-utils/src/utils/sandbox.js
@@ -29,10 +29,11 @@ import StyleSheet from '@instructure/ui-stylesheet'
 import ReactComponentWrapper from './reactComponentWrapper'
 
 import initConsole from './initConsole'
-import { newMutationObserver } from './shims'
+import { initGlobals, cleanupGlobals } from './globals'
 
 class Sandbox {
   constructor () {
+    initGlobals()
     initConsole()
 
     // eslint-disable-next-line no-console
@@ -46,7 +47,7 @@ class Sandbox {
     }
 
     this._addedNodes = []
-    this._observer = newMutationObserver((mutations) => {
+    this._observer = new MutationObserver((mutations) => {
       mutations.forEach((mutation) => {
         Array.from(mutation.addedNodes).forEach((addedNode) => {
           this._addedNodes.push(addedNode)
@@ -70,8 +71,8 @@ class Sandbox {
     this._sandbox.resetHistory()
     this._sandbox.restore()
 
-    window.localStorage.clear()
-    window.sessionStorage.clear()
+    window.localStorage && window.localStorage.clear()
+    window.sessionStorage &&  window.sessionStorage.clear()
 
     if (this._originalWindowOnError) {
       window.onerror = this._originalWindowOnError
@@ -92,11 +93,15 @@ class Sandbox {
     if (global.viewport) {
       global.viewport.set('large')
     }
+
+    cleanupGlobals()
   }
 
   setup () {
     try {
       this.teardown()
+
+      initGlobals()
 
       document.documentElement.setAttribute('dir', 'ltr')
       document.documentElement.setAttribute('lang', 'en-US')
